% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constant.R
\name{ConstantMeasure}
\alias{ConstantMeasure}
\title{Constant Measure}
\usage{
ConstantMeasure(constant = 0, ...)
}
\arguments{
\item{constant}{a non-negative numeric constant}
}
\description{
A trivial string measure that returns a constant for any pair of values.
}
