% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppMeasure.R
\docType{class}
\name{CppMeasure-class}
\alias{CppMeasure-class}
\title{Virtual Class for a Measure with a C++ Implementation}
\description{
This class serves as a trait which is possessed by measures that have a
C++ implementation. Measures without this trait are implemented in R,
and may be slower to execute.
}
