% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppSeqComparator.R
\docType{class}
\name{CppSeqComparator-class}
\alias{CppSeqComparator-class}
\title{Virtual Class for a Sequence Comparator with a C++ Implementation}
\description{
This class is a trait possessed by SequenceComparators that have a
C++ implementation. SequenceComparators without this trait are implemented
in R, and may be slower to execute.
}
