% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.R
\name{CustomMeasure}
\alias{CustomMeasure}
\title{Custom Measure}
\usage{
CustomMeasure(
  measure,
  symmetric = FALSE,
  distance = FALSE,
  similarity = FALSE,
  tri_inequal = FALSE,
  ...
)
}
\arguments{
\item{measure}{a vectorized function that takes a pair of vectors x and y
and returns a vector of elementwise distances/similarities.}

\item{symmetric}{whether the measure is symmetric if the pair of vectors
x and y are interchanged.}

\item{distance}{whether the measure can be interpreted as a
distance. Defaults to FALSE.}

\item{similarity}{whether the measure can be interpreted as a similarity.
Defaults to FALSE.}

\item{tri_inequal}{whether the measure satisfies the triangle inequality.
Defaults to FALSE.}
}
\description{
TODO
}
