% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamerauLevenshtein.R
\name{DamerauLevenshtein}
\alias{DamerauLevenshtein}
\title{Damerau-Levenshtein Distance}
\usage{
DamerauLevenshtein(
  deletion = 1,
  insertion = 1,
  substitution = 1,
  transposition = 1,
  normalize = FALSE,
  similarity = FALSE,
  ignore_case = FALSE,
  use_bytes = FALSE,
  ...
)
}
\arguments{
\item{deletion}{positive cost associated with deletion of a character.
Defaults to unit cost.}

\item{insertion}{positive cost associated insertion of a character.
Defaults to unit cost.}

\item{substitution}{positive cost associated with substitution of a
character. Defaults to unit cost.}

\item{transposition}{positive cost associated with transposing (swapping)
a pair of characters. Defaults to unit cost.}

\item{normalize}{a logical. If TRUE, distances are normalized to the
unit interval. Defaults to FALSE.}

\item{similarity}{a logical. If TRUE, similarity scores on the unit interval
are returned instead of distances. Defaults to FALSE.}

\item{ignore_case}{a logical. If TRUE, case is ignored when computing the
distance. Defaults to FALSE.}

\item{use_bytes}{a logical. If TRUE, distances are computed byte-by-byte
rather than character-by-character.}
}
\description{
The Damerau-Levenshtein distance between two strings is the minimum cost of
single-character operations (insertions, deletions, substitutions or
transpositions) required to transform one string into the other. It differs
from the Levenshtein distance by adding \emph{transpositions} (swaps) among the
allowable operations.
}
\note{
If the costs of all operations are identical, Damerau-Levenshtein
distance is a proper distance metric (it satisfies the triangle
inequality).
}
