% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\name{elementwise}
\alias{elementwise}
\alias{elementwise,Measure-method}
\title{Elementwise Similarity/Distance Vector}
\usage{
elementwise(measure, x, y, ...)

\S4method{elementwise}{Measure}(measure, x, y, ...)
}
\arguments{
\item{measure}{a measure used to compare the objects, which is a sub-class
of \code{\linkS4class{Measure}}.}

\item{x, y}{a collection of objects to compare, typically stored as entries
in an atomic vector, rows in a matrix, or entries in a list. The required
format depends on the type of \code{measure}. If \code{x} and \code{y} do not contain
the same number of objects, the smaller collection is recycled according
to standard \code{R} behavior.}

\item{...}{other parameters passed on to other methods.}
}
\value{
Every object in \code{x} is compared to every object in \code{y} elementwise
(with recycling) using the given measure , to produce a numeric vector of
scores of length \eqn{max{size(x), size(y)}}.
}
\description{
Computes elementwise similarities/distances between two collections of
objects (strings, vectors, etc.) using the provided measure.
}
\section{Methods (by class)}{
\itemize{
\item \code{Measure}: Calls \code{measure(x, y)} for any \code{x} and \code{y}
}}

\note{
This function is not strictly necessary, as the \code{measure} itself is a
function that returns elementwise vectors of scores. In other words,
\code{measure(x, y)} is equivalent to \code{elementwise(measure, x, y)}.
}
\examples{
## Compute the absolute difference between two sets of scalar observations
data("iris")
x <- as.matrix(iris$Sepal.Width)
y <- as.matrix(iris$Sepal.Length)
elementwise(Euclidean(), x, y)

## Compute the edit distance between columns of two linked data.frames
col.1 <- c("Hasna Yuhanna", "Korina Zenovia", "Phyllis Haywood", "Nicky Ellen")
col.2 <- c("Hasna Yuhanna", "Corinna Zenovia", "Phyllis Dorothy Haywood", "Nicole Ellen")
elementwise(Levenshtein(), col.1, col.2)
Levenshtein()(col.1, col.2)               # equivalent to above

## Recycling is used if the two collections don't contain the same number of objects
elementwise(Levenshtein(), "Cora Zenovia", col.1)

}
