% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hamming.R
\name{Hamming}
\alias{Hamming}
\title{Hamming Distance}
\usage{
Hamming(
  normalize = FALSE,
  similarity = FALSE,
  ignore_case = FALSE,
  use_bytes = FALSE,
  ...
)
}
\arguments{
\item{normalize}{a logical. If TRUE, distances are normalized to the
unit interval. Defaults to FALSE.}

\item{similarity}{a logical. If TRUE, similarity scores on the unit interval
are returned instead of distances. Defaults to FALSE.}

\item{ignore_case}{a logical. If TRUE, case is ignored when computing the
distance. Defaults to FALSE.}

\item{use_bytes}{a logical. If TRUE, the measure is computed byte-by-byte
rather than character-by-character.}
}
\description{
TODO
}
\note{
Unlike other edit distances, this is not a metric if normalize is TRUE
}
