% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lookup.R
\name{Lookup}
\alias{Lookup}
\title{Lookup Measure}
\usage{
Lookup(
  lookup_table,
  values_colnames,
  measure_colname,
  default_match = 0,
  default_nonmatch = Inf,
  symmetric = TRUE,
  ignore_case = FALSE,
  ...
)
}
\arguments{
\item{lookup_table}{data frame containing distances/similarities for
pairs of values}

\item{values_colnames}{character vector containing the colnames
corresponding to pairs of values (e.g. strings) in \code{lookup_table}}

\item{measure_colname}{name of column that contains distances/similarities
in \code{lookup_table}}

\item{default_match}{distance/similarity to use if the pair of values
match exactly and do not appear in \code{lookup_table}. Defaults to 0.0.}

\item{default_nonmatch}{distance/similarity to use if the pair of values are
not an exact match and do not appear in \verb{lookup table}. Defaults to Inf.}

\item{symmetric}{whether the underlying distance/similarity measure is
symmetric. If TRUE \code{lookup_table} need only contain entries for
one of the two pairs---i.e. an entry for value pair (y, x) is not required
if an entry for (x, y) is already present.}

\item{ignore_case}{a logical. If TRUE, case is ignored when computing the
distance/similarity.}
}
\description{
Computes a distances/similarities between pairs of values by retrieving them
from a provided lookup table.
}
