% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\docType{class}
\name{Measure-class}
\alias{Measure-class}
\title{Virtual Measure Class}
\description{
This class represents a measure for comparing pairs of objects.
It is the base class from which other types of measures (e.g.
\code{\linkS4class{NumericMeasure}} and \code{\linkS4class{StringMeasure}}) are derived.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a function which takes a pair of arguments \code{x} and \code{y}, and
returns the elementwise measure scores.}

\item{\code{symmetric}}{a logical of length 1. If TRUE, the measure is symmetric
in its arguments---i.e. \code{measure(x, y)} is identical to \code{measure(y, x)}.}

\item{\code{distance}}{a logical of length 1. If \code{TRUE}, the measure produces
distances and satisfies \code{measure(x, x) = 0}. The measure may not satisfy
all of the properties of a distance metric.}

\item{\code{similarity}}{a logical of length 1. If \code{TRUE}, the measure produces
similarity scores.}

\item{\code{tri_inequal}}{a logical of length 1. If \code{TRUE}, the measure satisfies
the triangle inequality. This is only possible (but not guaranteed) if
\code{distance = TRUE} and \code{symmetric = TRUE}.}
}}

