% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minkowski.R
\name{Minkowski}
\alias{Minkowski}
\title{Minkowski Distance}
\usage{
Minkowski(p = 2, ...)
}
\arguments{
\item{p}{TODO}
}
\description{
Compares a pair of strings based on whether they are in exact agreement or
not. Behaves as a distance measure by default.
}
