% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongeElkan.R
\name{MongeElkan}
\alias{MongeElkan}
\title{Monge-Elkan Measure}
\usage{
MongeElkan(
  inner_measure = Levenshtein(similarity = TRUE, normalize = TRUE),
  separator = "\\\\s+",
  agg_function = base::mean,
  symmetrize = FALSE,
  ...
)
}
\arguments{
\item{inner_measure}{inner string measure of class \code{\linkS4class{StringMeasure}}.
Defaults to \code{\link{Levenshtein}} similarity.}

\item{separator}{separator for tokens/words. Defaults to whitespace.}

\item{agg_function}{aggregation function to use when aggregating inner
similarities/distances between tokens. Defaults to arithmetic mean.
Harmonic mean may be a better choice for normalized similarity measures.}

\item{symmetrize}{logical indicating whether to use a symmetrized version
of the Monge-Elkan measure. Defaults to FALSE.}
}
\description{
TODO A hybrid string measure. Token- and character-based.
}
\details{
TODO
\deqn{\mathrm{sim_MongeElkan}(v_1, v_2) = \frac{1}{|v_1|} \sum_{i = 1}^{|v_1|} \max_{j} \mathrm{sim.inner}(v_{1,i}, v_{2,j})}
{sim_MongeElkan(v1,v2) = 1/|v1| Σ_i{ max_j sim.inner(v1_i, v2_j) }}
}
\references{
Monge, A. E., & Elkan, C. (1996, August). The Field Matching
Problem: Algorithms and Applications. In \emph{KDD} (Vol. 2, pp. 267-270).
}
