% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppMeasure.R
\name{pairwise,CppMeasure,vector,vector-method}
\alias{pairwise,CppMeasure,vector,vector-method}
\title{Pairwise}
\usage{
\S4method{pairwise}{CppMeasure,vector,vector}(measure, x, y, return_matrix = FALSE, ...)
}
\description{
Pairwise
}
