% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairwiseMatrix.R
\docType{class}
\name{PairwiseMatrix-class}
\alias{PairwiseMatrix-class}
\alias{as.PairwiseMatrix}
\alias{as.PairwiseMatrix,matrix-method}
\alias{as.PairwiseMatrix,numeric-method}
\alias{as.matrix,PairwiseMatrix-method}
\title{Pairwise Similarity/Distance Matrix}
\usage{
as.PairwiseMatrix(.Data, ...)

\S4method{as.PairwiseMatrix}{matrix}(.Data, ...)

\S4method{as.PairwiseMatrix}{numeric}(.Data, Dim, Diag, ...)

\S4method{as.matrix}{PairwiseMatrix}(x, ...)
}
\description{
Represents a pairwise similarity or distance matrix.
}
\details{
If the elements being compared are from the same set, the matrix
may be symmetric if the corresponding similarity/distance measure is
symmetric. In this case, entries in the upper triangle and/or along the
diagonal may not be stored in memory, since they are redundant.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{entries of the matrix in column-major order. Entries in the
upper triangle and/or on the diagonal may be omitted.}

\item{\code{Dim}}{integer vector of length 2. The dimensions of the matrix.}

\item{\code{Diag}}{logical indicating whether the diagonal entries are stored in
\code{.Data}.}
}}

