% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceComparator.R
\docType{class}
\name{SequenceComparator-class}
\alias{SequenceComparator-class}
\title{Virtual Sequence Comparator Class}
\description{
Represents a comparator for pairs of sequences.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a function that calls the elementwise method for this class,
with arguments \code{x}, \code{y} and \code{...}.}

\item{\code{symmetric}}{a logical of length 1. If TRUE, the comparator is symmetric
in its arguments---i.e. \code{comparator(x, y)} is identical to
\code{comparator(y, x)}.}

\item{\code{distance}}{a logical of length 1. If \code{TRUE}, the comparator produces
distances and satisfies \code{comparator(x, x) = 0}. The comparator may not
satisfy all of the properties of a distance metric.}

\item{\code{similarity}}{a logical of length 1. If \code{TRUE}, the comparator produces
similarity scores.}

\item{\code{tri_inequal}}{a logical of length 1. If \code{TRUE}, the comparator satisfies
the triangle inequality. This is only possible (but not guaranteed) if
\code{distance = TRUE} and \code{symmetric = TRUE}.}
}}

