% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringMeasure.R
\docType{class}
\name{StringMeasure-class}
\alias{StringMeasure-class}
\title{Virtual String Measure Class}
\description{
This class represents a measure for comparing pairs of strings.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a function which takes a pair of arguments \code{x} and \code{y}, and
returns the elementwise measure scores.}

\item{\code{symmetric}}{a logical of length 1. If TRUE, the measure is symmetric
in its arguments---i.e. \code{measure(x, y)} is identical to \code{measure(y, x)}.}

\item{\code{distance}}{a logical of length 1. If \code{TRUE}, the measure produces
distances and satisfies \code{measure(x, x) = 0}. The measure may not satisfy
all of the properties of a distance metric.}

\item{\code{similarity}}{a logical of length 1. If \code{TRUE}, the measure produces
similarity scores.}

\item{\code{tri_inequal}}{a logical of length 1. If \code{TRUE}, the measure satisfies
the triangle inequality. This is only possible (but not guaranteed) if
\code{distance = TRUE} and \code{symmetric = TRUE}.}

\item{\code{ignore_case}}{a logical of length 1. If TRUE, case is ignored when
comparing strings. Defaults to FALSE.}

\item{\code{use_bytes}}{a logical of length 1. If TRUE, strings are compared
byte-by-byte rather than character-by-character.}
}}

