% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryComp.R
\name{BinaryComp}
\alias{BinaryComp}
\title{Binary Comparison Measure}
\usage{
BinaryComp(disagree = 1, similarity = FALSE, ignore_case = FALSE, ...)
}
\arguments{
\item{disagree}{positive distance to return if the pair of strings are not
identical. Defaults to 1.0.}

\item{similarity}{a logical. If TRUE, similarities are returned instead of
distances. Specifically \code{disagree} is returned if the strings agree,
and 0.0 is returned otherwise.}

\item{ignore_case}{a logical. If TRUE, case is ignored when comparing the
strings.}
}
\description{
Compares a pair of strings based on whether they are in exact agreement or
not. Behaves as a distance measure by default.
}
