% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strcompr.R
\docType{package}
\name{comparator-package}
\alias{comparator}
\alias{comparator-package}
\title{comparator: Similarity and Distance Measures in R}
\description{
Provides similarity and distance measures for clustering and 
    record linkage applications.
}
\author{
\strong{Maintainer}: Neil Marchant \email{ngmarchant@gmail.com}

}
\keyword{internal}
