% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constant.R
\name{Constant}
\alias{Constant}
\title{Constant String/Sequence Comparator}
\usage{
Constant(constant = 0)
}
\arguments{
\item{constant}{a non-negative numeric vector of length 1. Defaults to zero.}
}
\value{
A \code{Constant} instance is returned, which is an S4 class inheriting
from \code{\linkS4class{StringComparator}}.
}
\description{
A trivial comparator that returns a constant for any pair of strings or
sequences.
}
