% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongeElkan.R
\name{MongeElkan}
\alias{MongeElkan}
\title{Monge-Elkan Measure}
\usage{
MongeElkan(
  inner_measure = Levenshtein(similarity = TRUE, normalize = TRUE),
  separator = "\\\\s+",
  agg_function = base::mean,
  symmetrize = FALSE
)
}
\arguments{
\item{inner_measure}{internal string measure of class
\code{\linkS4class{StringMeasure}}. Defaults to \code{\link{Levenshtein}} similarity.}

\item{separator}{separator for tokens/words. Defaults to whitespace.}

\item{agg_function}{aggregation function to use when aggregating internal
similarities/distances between tokens. Defaults to \code{\link{mean}},
however \code{\link{hmean}} may be a better choice for normalized similarity
measures.}

\item{symmetrize}{logical indicating whether to use a symmetrized version
of the Monge-Elkan measure. Defaults to FALSE.}
}
\value{
A \code{MongeElkan} instance is returned, which is an S4 class inheriting from
\code{\linkS4class{StringMeasure}}.
}
\description{
The Monge-Elkan measure is a hybrid token-character measure for comparing
a pair of multi-token (multi-word) strings \eqn{x} and \eqn{y}. It uses an
internal character-based measure to compute similarity scores between pairs
of tokens in \eqn{x} and \eqn{y}, then takes the average of the maximum
score for each token in \eqn{x}.
}
\details{
Given two token vectors \eqn{x} and \eqn{y}, the Monge-Elkan measure is
defined as:
\deqn{\mathrm{ME}(x, y) = \frac{1}{|x|} \sum_{i = 1}^{|x|} \max_j \mathrm{sim}(x_i, y_j)}{ME(x, y) = 1/|x| sum_i{ max_j sim(x_i, y_j) }}
where \eqn{x_i} is the i-th token (string) in \eqn{x}, \eqn{|x|} is the
number of tokens in \eqn{x} and \eqn{\mathrm{sim}}{sim} is an internal
character-based similarity measure.

A generalization of the original Monge-Elkan measure is implemented here,
which allows for internal distance measures in place of similarity measures,
and/or more general aggregation functions in place of the arithmetic mean.
The generalized Monge-Elkan measure is defined as:
\deqn{\mathrm{ME}(x, y) = \mathrm{agg}(\mathrm{opt}_j \ \mathrm{inner}(x_i, y_j))}{ME(x, y) = agg(opt_j inner(x_i, y_j))}
where \eqn{\mathrm{inner}}{inner} is an internal distance or similarity
measure, \eqn{\mathrm{opt}}{opt} is \eqn{\max}{max} if
\eqn{\mathrm{inner}}{inner} is a similarity measure or \eqn{\min}{min} if
it is a distance measure, and \eqn{\mathrm{agg}}{agg} is an aggregation
function which takes a vector of scores for each token in \eqn{x} and
returns a scalar.

By default, the Monge-Elkan measure is asymmetric in its arguments \eqn{x}
and \eqn{y}. If \code{symmetrize = TRUE}, a symmetric version of the measure
is obtained as follows
\deqn{\mathrm{ME}_{sym}(x, y) = \mathrm{opt} \ \{\mathrm{ME}(x, y), \mathrm{ME}(y, x)\}}{ME_sym(x, y) = opt { ME(x, y), ME(y, x) }}
where \eqn{\mathrm{opt}}{opt} is defined above.
}
\examples{
## Compare names with heterogenous representations
x <- "The University of California - San Diego"
y <- "Univ. Calif. San Diego"
MongeElkan()(x, y)

## The symmetrized variant is arguably more appropriate for this example
MongeElkan(symmetrize = TRUE)(x, y) 

## Using a different internal measure changes the result
MongeElkan(inner_measure = BinaryComp(), symmetrize=TRUE)(x, y)

}
\references{
Monge, A. E., & Elkan, C. (1996), "The Field Matching
Problem: Algorithms and Applications", In \emph{Proceedings of the Second
International Conference on Knowledge Discovery and Data Mining (KDD'96)},
pp. 267-270.

Jimenez, S., Becerra, C., Gelbukh, A., & Gonzalez, F. (2009), "Generalized
Monge-Elkan Method for Approximate Text String Comparison", In
\emph{Computational Linguistics and Intelligent Text Processing},
pp. 559-570.
}
